requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
decorator==4.4.2
environs==9.5.0
h11==0.13.0
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
marshmallow==3.17.0
MouseInfo==0.1.3
moviepy==1.0.3
mss==6.1.0
numpy==2.0.1
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
Pillow==10.4.0
proglog==0.1.10
PyAudio==0.2.14
PyAutoGUI==0.9.53
pycparser==2.21
pygame==2.6.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dotenv==0.20.0
pytweening==1.0.4
requests==2.28.1
robotframework==5.0.0
robotframework-pythonlibcore==3.0.0
robotframework-screencaplibrary==1.5.1
robotframework-seleniumlibrary==5.1.3
selenium==4.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
unicode==2.9
Unidecode==1.3.4
urllib3==1.26.11
wsproto==1.2.0